# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Factur-IA API",
    "version": "1.5.0"
  },
  "paths": {
...
    "/utilisateurs/me": {
      "get": {
        "tags": [
          "Gestion des Utilisateurs"
        ],
        "summary": "Get My Profile",
        "description": "Récupère le profil de l'utilisateur actuellement connecté.",
        "operationId": "get_my_profile_utilisateurs_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Gestion des Utilisateurs"
        ],
        "summary": "Update My Profile",
        "description": "Permet à l'utilisateur de modifier ses propres informations personnelles.",
        "operationId": "update_my_profile_utilisateurs_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtilisateurUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/utilisateurs/": {
      "get": {
        "tags": [
          "Gestion des Utilisateurs"
        ],
        "summary": "List Team Members",
        "description": "Liste tous les utilisateurs rattachés à l'entreprise active (l'équipe).",
        "operationId": "list_team_members_utilisateurs__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-entreprise-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "ID de l'entreprise",
              "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
            },
            "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UtilisateurRead"
                  },
                  "title": "Response List Team Members Utilisateurs  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Gestion des Utilisateurs"
        ],
        "summary": "Create Team Member",
        "description": "Crée un nouvel utilisateur et le rattache automatiquement à l'entreprise actuelle.",
        "operationId": "create_team_member_utilisateurs__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-entreprise-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "ID de l'entreprise",
              "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
            },
            "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtilisateurCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/utilisateurs/inscription": {
      "post": {
        "tags": [
          "Gestion des Utilisateurs"
        ],
        "summary": "Register Public User",
        "description": "Inscription publique d'un nouvel utilisateur.\nCrée un compte utilisateur \"nu\", sans rattachement initial à une entreprise.",
        "operationId": "register_public_user_utilisateurs_inscription_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtilisateurCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/utilisateurs/{user_id}": {
      "delete": {
        "tags": [
          "Gestion des Utilisateurs"
        ],
        "summary": "Delete Team Member",
        "operationId": "delete_team_member_utilisateurs__user_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "x-entreprise-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "ID de l'entreprise",
              "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
            },
            "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Gestion des Utilisateurs"
        ],
        "summary": "Update Team Member",
        "description": "Modifie les informations d'un\ncollaborateur (incluant ses rôles et accès admin).",
        "operationId": "update_team_member_utilisateurs__user_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "x-entreprise-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "ID de l'entreprise",
              "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
            },
            "description": "Identifiant de l'entreprise (tenant)                 actif transmis dans les en-têtes."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtilisateurTeamUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UtilisateurCreate": {
        "properties": {
          "nom": {
            "type": "string",
            "maxLength": 255,
            "title": "Nom"
          },
          "prenom": {
            "type": "string",
            "maxLength": 255,
            "title": "Prenom"
          },
          "adresse": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Adresse"
          },
          "adresse_complement": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Adresse Complement"
          },
          "code_postal": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Code Postal"
          },
          "ville": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Ville"
          },
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email",
            "title": "Email"
          },
          "telephone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Telephone"
          },
          "est_actif": {
            "type": "boolean",
            "title": "Est Actif",
            "default": true
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "Mot de passe en clair"
          },
          "id_role": {
            "type": "integer",
            "title": "Id Role",
            "description": "ID du rôle métier rattaché"
          },
          "est_admin": {
            "type": "boolean",
            "title": "Est Admin",
            "description": "Droit admin au niveau de l'entreprise",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "nom",
          "prenom",
          "email",
          "password",
          "id_role"
        ],
        "title": "UtilisateurCreate",
        "description": "Schéma pour créer un utilisateur avec son mot de passe en clair."
      },
      "UtilisateurRead": {
        "properties": {
          "nom": {
            "type": "string",
            "maxLength": 255,
            "title": "Nom"
          },
          "prenom": {
            "type": "string",
            "maxLength": 255,
            "title": "Prenom"
          },
          "adresse": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Adresse"
          },
          "adresse_complement": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Adresse Complement"
          },
          "code_postal": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Code Postal"
          },
          "ville": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Ville"
          },
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email",
            "title": "Email"
          },
          "telephone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Telephone"
          },
          "est_actif": {
            "type": "boolean",
            "title": "Est Actif",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "date_creation": {
            "type": "string",
            "format": "date-time",
            "title": "Date Creation"
          },
          "date_modification": {
            "type": "string",
            "format": "date-time",
            "title": "Date Modification"
          },
          "date_derniere_connexion": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Derniere Connexion"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "est_admin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Est Admin",
            "description": "Statut administrateur du membre dans l'entreprise active. Lecture seule : à utiliser pour pré-remplir le formulaire d'édition et éviter de retirer les droits par erreur."
          }
        },
        "type": "object",
        "required": [
          "nom",
          "prenom",
          "email",
          "id",
          "date_creation",
          "date_modification"
        ],
        "title": "UtilisateurRead",
        "description": "Schéma de sortie (le hash n'est jamais renvoyé)."
      },
      "UtilisateurTeamUpdate": {
        "properties": {
          "nom": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nom"
          },
          "prenom": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prenom"
          },
          "adresse": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adresse"
          },
          "adresse_complement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adresse Complement"
          },
          "code_postal": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Code Postal"
          },
          "ville": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Ville"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telephone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telephone"
          },
          "est_actif": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Est Actif"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "id_role": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Role"
          },
          "est_admin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Est Admin"
          }
        },
        "type": "object",
        "title": "UtilisateurTeamUpdate",
        "description": "Hérite de tous les champs optionnels (nom, prenom, adresse...) de UtilisateurUpdate,\net y ajoute les champs de gestion réservés aux administrateurs."
      },
      "UtilisateurUpdate": {
        "properties": {
          "nom": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nom"
          },
          "prenom": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prenom"
          },
          "adresse": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adresse"
          },
          "adresse_complement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adresse Complement"
          },
          "code_postal": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Code Postal"
          },
          "ville": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Ville"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telephone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telephone"
          },
          "est_actif": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Est Actif"
          }
        },
        "type": "object",
        "title": "UtilisateurUpdate",
        "description": "Schéma pour la mise à jour partielle d'un utilisateur."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-OCR-Secret-Token"
      }
    }
  }
}